+++ lib/twitch/twitch_token_store.dart
@@ -8,7 +8,7 @@ abstract interface class TwitchTokenStore {
 }
 
 final class SharedPreferencesTwitchTokenStore implements TwitchTokenStore {
-  static const String tokenKey = 'twitch_creds';
+  static const String tokenKey = 'twitch_oauth';
 
   SharedPreferencesTwitchTokenStore([this._preferences]);